% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_DNA.seq}
\alias{demo_DNA.seq}
\title{A demo of DNA sequences}
\format{A list contains 10 DNA sequences.
\describe{
The sequences are loaded by function \code{\link[seqinr]{read.fasta}}.
}}
\source{
DNA sequences are selected from GENCODE.
}
\usage{
demo_DNA.seq
}
\description{
This file contains 10 DNA sequences.
}
\keyword{datasets}
