% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_SS.seq}
\alias{demo_SS.seq}
\title{A demo of secondary structure sequences}
\format{A data frame with 3 rows and 10 variables:
\describe{
The first row is RNA sequence; the second row is Dot-Bracket Notation of
secondary structure sequences; the last row is minimum free energy (MFE).
}}
\source{
DNA sequences are selected from GENCODE. Secondary structure of each
sequence is obtained from program "RNAfold".
}
\usage{
data(demo_SS.seq)
}
\description{
This file contains 10 SS (Secondary Structure) sequences.
}
